finish organizing modules
