+++ src/command.rs
@@ -1,4 +1,5 @@
 pub(crate) mod build;
 pub(crate) mod clean;
 pub(crate) mod list;
+pub(crate) mod new;
 pub(crate) mod run;
