+++ src/util.rs
@@ -1,63 +1,8 @@
-use std::{collections, fmt, fs, path};
+use std::{env, fmt, fs, io, path};
 
 pub(crate) use colored::Colorize;
 
 use crate::error::{Message, TomlErrorEmitter};
-
-/// Acceptable plugin config toml names for renaming and symlinking REAPER extenion plugins built with Rust.
-pub(crate) const CONFIG_FILE_NAMES: &[&str; 2] = &[".reaper.toml", "reaper.toml"];
-
-/// The dynamically linked C library Windows extension
-pub(crate) const WINDOWS_PLUGIN_EXT: &str = ".dll";
-/// The dynamically linked C library Linux extension
-pub(crate) const LINUX_PLUGIN_EXT: &str = ".so";
-/// The dynamically linked C library MacOS (Darwin) extension
-pub(crate) const DARWIN_PLUGIN_EXT: &str = ".dylib";
-
-/// The parsed contents of a `reaper.toml` config file.
-#[derive(Debug, serde::Deserialize)]
-pub(crate) struct ReaperPluginConfig {
-    /// The path to the `reaper.toml` config file.
-    #[serde(skip)]
-    file: path::PathBuf,
-
-    /// The contents of a `reaper.toml` config file as a [`std::str::String`].
-    #[serde(skip)]
-    contents: String,
-
-    /// The contents of a deserialized `reaper.toml` config file.
-    extension_plugins: collections::HashMap<toml::Spanned<String>, toml::Spanned<path::PathBuf>>,
-}
-impl ReaperPluginConfig {
-    /// The path to the `reaper.toml` config file.
-    pub(crate) fn file(&self) -> &path::PathBuf {
-        &self.file
-    }
-
-    /// The path to the `reaper.toml` config file.
-    pub(crate) fn contents(&self) -> &str {
-        &self.contents
-    }
-
-    /// Locate and deserialize a `reaper.toml` config file.
-    pub(crate) fn load(project_root: &path::Path) -> anyhow::Result<Self> {
-        let config_file = CONFIG_FILE_NAMES
-            .iter()
-            .map(|config_file_name| project_root.join(config_file_name))
-            .find(|config_path| config_path.exists())
-            .unwrap(); // We already ensured this path exists
-        let config_contents = fs::read_to_string(&config_file)
-            .map_err(|err| anyhow::anyhow!("failed to read reaper toml file:\n{err:#?}"))?;
-
-        let mut config: Self = toml::from_str(&config_contents).map_err(|err| {
-            anyhow::anyhow!("failed to load plugin config from reaper toml:\n{err:#?}")
-        })?;
-        config.file = config_file;
-        config.contents = config_contents;
-
-        Ok(config)
-    }
-}
 
 /// Represents a REAPER plugin's manifest information.
 #[derive(Debug, PartialEq, Eq, PartialOrd, Ord)]
@@ -93,6 +38,26 @@ impl fmt::Display for PluginManifest {
         }
         Ok(())
     }
+}
+
+pub(crate) fn find_project_root() -> anyhow::Result<path::PathBuf> {
+    let mut current_dir = env::current_dir()?;
+
+    loop {
+        if current_dir.join("Cargo.toml").is_file() && current_dir.join(".reaper.toml").is_file()
+            || current_dir.join("reaper.toml").is_file()
+        {
+            return Ok(current_dir);
+        }
+
+        if !current_dir.pop() {
+            break;
+        }
+    }
+
+    anyhow::bail!(
+        "Unable to find project root directory. Please ensure a reaper.toml or .reaper.toml file is present in the project root, and try again."
+    )
 }
 
 /// Processes the reaper config toml and the plugin `Cargo.toml` files, collecting diagnostic errors and returning the plugin's manifest.
@@ -177,4 +142,312 @@ pub(crate) fn validate_plugin(
         );
     }
     Ok(manifest)
+}
+
+/// Rename the resulting extension plugin, returning the new plugin path if it succeeds.
+///
+/// > Note: This function is platform agnostic
+///
+/// # Usage
+///
+/// This is run automatically when running the `cargo reaper` command.
+pub(crate) fn _rename_plugin(
+    project_root: &path::Path,
+    profile: &str,
+    old_plugin_path: &path::PathBuf,
+    plugin_name_to: &str,
+) -> anyhow::Result<path::PathBuf> {
+    let new_plugin_path = project_root
+        .join("target")
+        .join(profile)
+        .join(plugin_name_to);
+
+    fs::rename(old_plugin_path, &new_plugin_path)
+        .map_err(|err| anyhow::anyhow!("failed to rename plugin: {err:?}"))?;
+
+    println!(
+        "    {} plugin renamed {} → {}",
+        "Finished".green().bold(),
+        old_plugin_path.display(),
+        new_plugin_path.display()
+    );
+
+    Ok(new_plugin_path)
+}
+
+/// Symlink the REAPER extension plugin to the `UserPlugins` directory.
+///
+/// > Note: This function is platform agnostic
+///
+/// # Usage
+///
+/// This is run automatically when running the `cargo reaper build` command, unless passed `--no-symlink`.
+pub(crate) fn _symlink_plugin<S>(
+    plugin_path: &path::PathBuf,
+    user_plugins_dir: &path::Path,
+    symlink_plugin: S,
+) -> anyhow::Result<()>
+where
+    S: FnOnce(&path::PathBuf, &path::PathBuf) -> io::Result<()>,
+{
+    if !user_plugins_dir.exists() {
+        anyhow::bail!(
+            "The 'UserPlugins' directory must exist before the plugin can be symlinked. Please launch REAPER to initialize the 'UserPlugins' directory and try again."
+        );
+    }
+
+    let symlink_path = user_plugins_dir.join(plugin_path.file_name().ok_or_else(|| {
+        anyhow::anyhow!(
+            "Unable to get plugin file name from path '{}'",
+            plugin_path.display()
+        )
+    })?);
+    if symlink_path.exists() {
+        let currently_symlinked_plugin_path = fs::read_link(&symlink_path)?;
+        if &currently_symlinked_plugin_path != plugin_path {
+            println!(
+                "{}: removing stale symlink ({})",
+                "warning".yellow().bold(),
+                symlink_path.display()
+            );
+            fs::remove_file(&symlink_path)?;
+        } else {
+            println!(
+                "    {} symbolic link already exists ({})",
+                "Finished".green().bold(),
+                symlink_path.display(),
+            );
+            return Ok(());
+        }
+    }
+
+    // TODO: Sometimes this will still fail with 'AlreadyExists' errors. We should also go ahead and catch them here.
+    symlink_plugin(plugin_path, &symlink_path)
+        .map_err(|err| anyhow::anyhow!("failed to link extension plugin: {err:?}"))?;
+
+    println!(
+        "    {} symbolic link created {} → {}",
+        "Finished".green().bold(),
+        symlink_path.display(),
+        plugin_path.display()
+    );
+
+    Ok(())
+}
+
+/// Remove a REAPER extension plugin symlink from the `UserPlugins` directory.
+///
+/// > Note: This function is platform agnostic
+///
+/// # Usage
+///
+/// This is run automatically when running the `cargo reaper clean` command.
+pub(crate) fn _remove_plugin_symlink(
+    plugin_name: &str,
+    plugin_file_name: &str,
+    user_plugins_dir: &path::Path,
+    dry_run: bool,
+) -> anyhow::Result<()> {
+    let symlink_path = user_plugins_dir.join(plugin_file_name);
+    if symlink_path.is_symlink() {
+        if !dry_run {
+            fs::remove_file(&symlink_path).map_err(|err| {
+                anyhow::anyhow!("failed to remove symlink for `{plugin_name}`:\n{err:#?}")
+            })?;
+        }
+        return Ok(());
+    }
+
+    anyhow::bail!(
+        "`{}` does not contain a symlink for `{}` ({})",
+        user_plugins_dir.display(),
+        plugin_name,
+        plugin_file_name
+    )
+}
+
+#[cfg(target_os = "windows")]
+pub(crate) mod os {
+    //! Operating system specific functionality for handling operations which require knownledge of
+    //! either dynamic library file extensions, or interacting with the `UserPlugins` directory.
+
+    use std::{io, os, path};
+
+    use super::{_remove_plugin_symlink, _rename_plugin, _symlink_plugin};
+
+    /// The dynamically linked C library Windows extension
+    pub(crate) const WINDOWS_PLUGIN_EXT: &str = ".dll";
+
+    pub(crate) fn from_plugin_file_name(lib_name: &str) -> String {
+        lib_name.to_string()
+    }
+
+    pub(crate) fn add_plugin_ext(lib_name: &str) -> String {
+        format!("{lib_name}{WINDOWS_PLUGIN_EXT}")
+    }
+
+    pub(crate) fn rename_plugin(
+        project_root: &path::Path,
+        profile: &str,
+        old_plugin_path: &path::PathBuf,
+        plugin_name_to: &str,
+    ) -> anyhow::Result<path::PathBuf> {
+        _rename_plugin(project_root, profile, old_plugin_path, plugin_name_to)
+    }
+
+    pub(crate) fn symlink_plugin(plugin_path: &path::PathBuf) -> anyhow::Result<()> {
+        _symlink_plugin(
+            plugin_path,
+            &dirs::data_dir()
+                .ok_or_else(|| anyhow::anyhow!("Unable to find 'AppData' directory"))?
+                .join("REAPER")
+                .join("UserPlugins"),
+            |plugin_path, symlink_path| {
+                os::windows::fs::symlink_file(plugin_path, symlink_path).map_err(|err|
+                    if format!("{err:?}").contains("A required privilege is not held by the client.") {
+                        io::Error::new(
+                            io::ErrorKind::PermissionDenied,
+                            "Windows treats symlink creation as a privileged action, therefore this function is likely to fail unless the user makes changes to their system to permit symlink creation. Users can try enabling Developer Mode, granting the SeCreateSymbolicLinkPrivilege privilege, or running the process as an administrator.",
+                        )
+                    } else {
+                        err
+                    }
+                )
+            },
+        )
+    }
+
+    pub(crate) fn remove_plugin_symlink(
+        plugin_name: &str,
+        plugin_file_name: &str,
+        dry_run: bool,
+    ) -> anyhow::Result<()> {
+        _remove_plugin_symlink(
+            plugin_name,
+            plugin_file_name,
+            &dirs::data_dir()
+                .ok_or_else(|| anyhow::anyhow!("Unable to find 'AppData' directory"))?
+                .join("REAPER")
+                .join("UserPlugins"),
+            dry_run,
+        )
+    }
+}
+
+#[cfg(target_os = "linux")]
+pub(crate) mod os {
+    //! Operating system specific functionality for handling operations which require knownledge of
+    //! either dynamic library file extensions, or interacting with the `UserPlugins` directory.
+
+    use std::{os, path};
+
+    use super::{_remove_plugin_symlink, _rename_plugin, _symlink_plugin};
+
+    /// The dynamically linked C library Linux extension
+    pub(crate) const LINUX_PLUGIN_EXT: &str = ".so";
+
+    pub(crate) fn from_plugin_file_name(lib_name: &str) -> String {
+        format!("lib{lib_name}")
+    }
+
+    pub(crate) fn add_plugin_ext(lib_name: &str) -> String {
+        format!("{lib_name}{LINUX_PLUGIN_EXT}")
+    }
+
+    pub(crate) fn rename_plugin(
+        project_root: &path::Path,
+        profile: &str,
+        old_plugin_path: &path::PathBuf,
+        plugin_name_to: &str,
+    ) -> anyhow::Result<path::PathBuf> {
+        _rename_plugin(project_root, profile, old_plugin_path, plugin_name_to)
+    }
+
+    pub(crate) fn symlink_plugin(plugin_path: &path::PathBuf) -> anyhow::Result<()> {
+        _symlink_plugin(
+            plugin_path,
+            &dirs::config_dir()
+                .ok_or_else(|| anyhow::anyhow!("Unable to find '.config' directory"))?
+                .join("REAPER")
+                .join("UserPlugins"),
+            |plugin_path, symlink_path| os::unix::fs::symlink(plugin_path, symlink_path),
+        )
+    }
+
+    pub(crate) fn remove_plugin_symlink(
+        plugin_name: &str,
+        plugin_file_name: &str,
+        dry_run: bool,
+    ) -> anyhow::Result<()> {
+        _remove_plugin_symlink(
+            plugin_name,
+            plugin_file_name,
+            &dirs::config_dir()
+                .ok_or_else(|| anyhow::anyhow!("Unable to find '.config' directory"))?
+                .join("REAPER")
+                .join("UserPlugins"),
+            dry_run,
+        )
+    }
+}
+
+#[cfg(target_os = "macos")]
+pub(crate) mod os {
+    //! Operating system specific functionality for handling operations which require knownledge of
+    //! either dynamic library file extensions, or interacting with the `UserPlugins` directory.
+
+    use std::{os, path};
+
+    use super::{_remove_plugin_symlink, _rename_plugin, _symlink_plugin};
+
+    /// The dynamically linked C library MacOS (Darwin) extension
+    pub(crate) const DARWIN_PLUGIN_EXT: &str = ".dylib";
+
+    pub(crate) fn from_plugin_file_name(lib_name: &str) -> String {
+        format!("lib{lib_name}")
+    }
+
+    pub(crate) fn add_plugin_ext(lib_name: &str) -> String {
+        format!("{lib_name}{DARWIN_PLUGIN_EXT}")
+    }
+
+    pub(crate) fn rename_plugin(
+        project_root: &path::Path,
+        profile: &str,
+        old_plugin_path: &path::PathBuf,
+        plugin_name_to: &str,
+    ) -> anyhow::Result<path::PathBuf> {
+        _rename_plugin(project_root, profile, old_plugin_path, plugin_name_to)
+    }
+
+    pub(crate) fn symlink_plugin(plugin_path: &path::PathBuf) -> anyhow::Result<()> {
+        _symlink_plugin(
+            plugin_path,
+            &dirs::home_dir()
+                .ok_or_else(|| anyhow::anyhow!("Unable to find 'Users' directory"))?
+                .join("Library")
+                .join("Application Support")
+                .join("REAPER")
+                .join("UserPlugins"),
+            |plugin_path, symlink_path| os::unix::fs::symlink(plugin_path, symlink_path),
+        )
+    }
+
+    pub(crate) fn remove_plugin_symlink(
+        plugin_name: &str,
+        plugin_file_name: &str,
+        dry_run: bool,
+    ) -> anyhow::Result<()> {
+        _remove_plugin_symlink(
+            plugin_name,
+            plugin_file_name,
+            &dirs::home_dir()
+                .ok_or_else(|| anyhow::anyhow!("Unable to find 'Users' directory"))?
+                .join("Library")
+                .join("Application Support")
+                .join("REAPER")
+                .join("UserPlugins"),
+            dry_run,
+        )
+    }
 }
